use abstract_piece*** as the _board_ and add update_moves functions
update_moves will search all posible moves for the piece and add
them to the avail_moves list.

There is a special rule for general that needed to take care later,
which is two generals can't be in the same file without pieces
between them. Or the player who makes such move loses.